+++ micro/classes/__init__.py
@@ -1,7 +1,7 @@
 from .database import Database
-from .badWords import BadWords
+from .bad_words import BadWords
 from .github import GitHub
-from .QuestionQuery import QuestionQuery
+from .question_query import QuestionQuery
 from .resources import Resources
 from .colored_embed import CEmbed
