+++ micro/classes/bad_words.py
